requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.2.3            # via django
django-extensions==2.2.6
django==3.0.3
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3              # via django
six==1.14.0               # via django-extensions, python-dateutil
sqlparse==0.3.0           # via django

# The following packages are considered to be unsafe in a requirements file:
# setuptools
